Separate configuration and instanciation from implementation
